% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.dotplot.R
\name{lefse.dotplot}
\alias{lefse.dotplot}
\title{dot plot using the data from lefse.res.clean function based on ggplot2}
\usage{
lefse.dotplot(data, p.size = 3)
}
\arguments{
\item{data}{the input data from lefse.res.clean func}

\item{p.size}{the size of point}

\item{col}{the colour of point}

\item{legend.position}{the position of legend}

\item{axis.text.color}{the color of axis test}

\item{...}{other parameters}
}
\value{
a ggplot obj
}
\description{
dot plot using the data from lefse.res.clean function based on ggplot2
}
\author{
ZhonghuiGai
}
