% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.R
\name{lefse}
\alias{lefse}
\title{lefse analysis by using the data from lefse.rawdata function}
\usage{
lefse(input = "lefse.txt")
}
\arguments{
\item{input}{the input data for lefse analysis}
}
\value{
lefse analysi results and plot
}
\description{
lefse analysis by using the data from lefse.rawdata function
}
\examples{
lefse(input = "lefse.txt")
}
\author{
ZhonghuiGai
}
