% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse.res.clean.R
\name{lefse.res.clean}
\alias{lefse.res.clean}
\title{clean lefse analysis result for ggplot2}
\usage{
lefse.res.clean(lefse.res = "B_lefse.res")
}
\arguments{
\item{lefse.res}{the first result of lefse analysis using lefse function}
}
\value{
a data frame
}
\description{
clean lefse analysis result for ggplot2
}
\examples{
lefse.res.clean(lefse.res = "B_lefse.res")
}
\author{
ZhonghuiGai
}
