% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_barplot.R
\name{lefse_barplot}
\alias{lefse_barplot}
\title{Draw a bar plot using the lefse analysis results}
\usage{
lefse_barplot(data, size = 0.5, col = NULL, tpye = "npg")
}
\arguments{
\item{data}{a data frame containing markers, grouping and LDA_value information}

\item{size}{the width of bar}

\item{col}{the col of bars, default value if NULL}

\item{tpye}{the color type of gfun::scale_color_manual2}
}
\value{
a ggplot2 bar plot
}
\description{
Draw a bar plot using the lefse analysis results
}
\examples{
library(gfun)
lefse.genus <- data.in("g_LDA.xlsx", rowname = F)
colnames(lefse.genus)[2] <- "group"
lefse.genus$markers <- factor(lefse.genus$markers, levels = lefse.genus$markers)
lefse.genus$group <- factor(lefse.genus$group, levels = c("HC", "A", "B"), labels = c("HC", "No-NAs", "TDF"))
lefse_barplot(data = lefse.genus) +  ggtheme::theme_size()
}
\author{
Zhonghui Gai
}
