% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lefse_rawdata.R
\name{lefse_rawdata}
\alias{lefse_rawdata}
\title{Prepare data for downstream Lefse analysis using asv.table, taxa info and the grouping metadata}
\usage{
lefse_rawdata(asv.path, taxa.path, group.path, lefse = TRUE)
}
\arguments{
\item{asv.path}{the path of ASVs_table obtained using usearch}

\item{taxa.path}{the path of ASV_taxa info obtained suing usearch}

\item{group.path}{the grouping metadata}

\item{lefse}{the logical value for which data to save}
}
\value{
a txt text file for lefse analysis
}
\description{
Prepare data for downstream Lefse analysis using asv.table, taxa info and the grouping metadata
}
\examples{
lefse_rawdata(asv.path = "ASVs_norm.txt", taxa.path = "ASV_tax2.txt", group.path = "group.txt", lefse = TRUE)
}
\author{
ZHonghuiGai
}
