% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_wrap.R
\name{str_wrap}
\alias{str_wrap}
\title{str_wrap}
\usage{
str_wrap(string, width = getOption("width"))
}
\arguments{
\item{string}{input string to be wrapped}

\item{width}{the length characters of each line}
}
\value{
a wrapped characters
}
\description{
Wraping long string to multiple lines, adapted from yulab.utils
}
\examples{
x <- c("this is a   looooooooong text", "this is a shorter text")
str_wrap(string = x, width = 10)
}
\author{
Zhonghui Gai
}
